Export types outside the suggestion function
